backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-openai
langchain-groq
pinecone-client
pypdf
bleach
slowapi
python-multipart
requests